:bug: Fix minor bugs
