+++ services/blockchain-indexer/shared/utils/transactionsUtils.js
@@ -14,12 +14,23 @@
  *
  */
 const { parseToJSONCompatObj } = require('./parser');
+const { TRANSACTION_STATUS, EVENT } = require('../constants');
 
 const normalizeTransaction = async tx => {
 	tx.moduleCommand = `${tx.module}:${tx.command}`;
 	return parseToJSONCompatObj(tx);
 };
 
+const getTransactionExecutionStatus = (tx, events) => {
+	const expectedEventName = `${tx.module}:${EVENT.COMMAND_EXECUTION_RESULT}`;
+	const commandExecResultEvents = events.filter(e => `${e.module}:${e.name}` === expectedEventName);
+	const txExecResultEvent = commandExecResultEvents.find(e => e.topics.includes(tx.id));
+	if (!txExecResultEvent) throw Error(`Event unavailable to determine execution status for transaction: ${tx.id}.`);
+
+	return txExecResultEvent.data.success ? TRANSACTION_STATUS.SUCCESS : TRANSACTION_STATUS.FAIL;
+};
+
 module.exports = {
 	normalizeTransaction,
+	getTransactionExecutionStatus,
 };
